# Remove array destructuring in colorDelta, add colorDeltaChannels, deprecate colorDelta