I'm looking to improve performance and I observed that if we avoid array destructuring in this function, we can make my performance test that uses this function heavily go from ~10s to ~6.5s. This is surprising to me, but the tradeoff here is pretty obvious.

Avoiding the arrays entirely as in `colorDeltaChannels` improves my performance test from ~6.5s to ~6.2s.