@lencioni 2024-09-15
I did some profiling of this change across a bunch of images, and here's the before:

![image](https://github.com/user-attachments/assets/c1fcc232-891b-45fb-a22a-4562ce505ed2)

Here's the after: 
![image](https://github.com/user-attachments/assets/5f8d2cda-f7b2-4e5c-991d-3d83d62d3a54)

The colorDelta time disappears almost entirely and the imageDiff (different repo) time goes down a bunch since it doesn't need to create the arrays in the first place.